added SocialConfig
